requirements.txt
nextcord
jishaku==2.4.0
python-dotenv
